+++ src/index.js
@@ -1,6 +1,5 @@
-import "./style.css";
-import "bootstrap/dist/css/bootstrap.min.css";
-import Nav from "./modules/Nav";
-import {getLikes} from './modules/Likes';
-getLikes();
+import './style.css';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import Nav from './modules/Nav';
+
 Nav();
